pyproject.toml
[project]
name = "new-folder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "datasets>=4.4.2",
    "fastapi>=0.128.0",
    "langchain>=1.2.2",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "langchain-text-splitters>=1.1.0",
    "ollama>=0.6.1",
    "pinecone>=8.0.0",
    "pypdf>=6.5.0",
    "python-dotenv>=1.2.1",
    "ragas>=0.4.2",
    "streamlit>=1.52.2",
    "uvicorn[standard]>=0.40.0",
]
